Change DioFactory base URL to production
Switch the base URL in DioFactory from the development environment to the production environment by updating the baseUrl parameter in the BaseOptions configuration to use AppLinks.baseUrlProd instead of AppLinks.baseUrlDev. This ensures that network requests made through the Dio instance are directed to the production server, aligning the application's behavior with its intended operational environment.

The change affects the initialization of the Dio instance within the DioFactory class, specifically targeting the configuration of the BaseOptions object, which is used to set the default settings for all requests made through the Dio instance. This update is crucial for maintaining the correct endpoint connectivity during the application's deployment phase.