✨🎨 feat(distribute_students): Improve layout and alignment
The changes made to the `distribute_students.dart` file focus on improving the layout and alignment of the elements within the `DistributeStudents` widget. The main changes are:

- Added `MainAxisAlignment.end` and `CrossAxisAlignment.start` to the
  column containing the blocked class desks to align them to the
  bottom-left corner.
- Removed the `width: double.infinity` property from the `Container`
  containing the blocked class desks, as it was not necessary.

These changes enhance the visual presentation and organization of the
distribute students screen, making it more intuitive and visually
appealing for the users.