✨ feat(control-system): Update API endpoint
Updating the API endpoint from `http://localhost:3333/` to `https://control-o5xlbifnea-uc.a.run.app/` in the `main.dart.js` file. This change is necessary to
connect the application to the production API hosted on Google Cloud Run.

Additionally, the `flutter_bootstrap.js` file is updated to include the new service worker version.