🚀 feat: Implement role screens expansion
This change introduces the use of `ExpansionTile` in the `RolesScreen` to display the list of screens associated with each role. Previously, the screens were displayed inline, but this change allows for a more organized and expandable view. The change also includes handling the case where no screens are found for a particular role.