✨ feat(auth): update API endpoint
The changes made in this commit update the API endpoint used in the `main.dart.js` file. The previous endpoint `http://localhost:3333/` has been replaced with `https://control-o5xlbifnea-uc.a.run.app/`, which is the new production API endpoint.

This change is necessary to ensure that the application connects to the correct API server in the production environment, as the local development endpoint is no longer valid.