Add User School Management Feature
This commit introduces a new feature for managing user-school associations in the system. It includes the addition of a new model, updates to existing models, and modifications to controllers and views to support the feature.

- Added `UserHasSchoolResModel` to handle user-school relationships.
- Updated `UserResModel` to include `UserHasSchoolResModel` for managing school associations.
- Modified `AdminController` to handle fetching and editing user-school relationships.
- Added new API endpoints in `app_links.dart` for managing user-school associations.
- Updated `app_constatnts.dart` to include a new role type, 'Super Admin'.
- Added a new widget, `AddSchoolUser`, to the admin screen for managing user-school assignments.
- Updated `all_user_widget.dart` to include a button for managing user-school associations.

This feature enhances the system's capability to manage user permissions and associations with schools more effectively.