✨ feat(firebase): add cache control headers
Adds cache control headers to the firebase hosting configuration. This will
ensure that HTML files are not cached and that other static assets have a
max-age of 0 and must be revalidated on each request. This will improve the
user experience by ensuring that users always see the latest version of the
application.