✨💻 feat(web): improve Dart WASM support
Adds support for Dart WASM compilation target in the Flutter web
loader. This includes the ability to load and initialize a
WebAssembly-based Flutter web app.

The changes focus on the following improvements:

- Added support for loading Dart WASM-based entrypoints
- Optimized the script loading process for Dart WASM apps
- Integrated Trusted Types policy support for script URLs
- Maintained backwards compatibility with existing JS-based
  entrypoints

These changes will enable developers to easily deploy Flutter web
apps that leverage the performance and features of WebAssembly.